add secret tests
